[Refactor] Controller에서 Domain모델을 사용하던 부분을 DTO로 변경

+++ src/main/java/com/yenso/yensoserver/Domain/Model/Job.java
@@ -12,7 +12,7 @@ public class Job implements Serializable {
 
     @Id
     @GeneratedValue(strategy = GenerationType.AUTO)
-    private int job_id;
+    private Long job_id;
 
     @Column(unique = true)
     private String category;
